% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{build_time_objects}
\alias{build_time_objects}
\title{Build time objects for interpolation}
\usage{
build_time_objects(periods_starts, periods_ends)
}
\arguments{
\item{periods_starts}{Vector of periods starts}

\item{periods_ends}{Vector of periods ends}
}
\value{
A list with the timing and the number of weeks
}
\description{
Build time objects for interpolation
}
