% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay_functions.R
\name{decay_functions}
\alias{decay_functions}
\alias{logistic_decay}
\alias{logistic_simple_decay}
\alias{exponential_decay}
\title{List of decay functions}
\usage{
decay_functions()

logistic_decay()

logistic_simple_decay()

exponential_decay()
}
\value{
List of decay functions

Logistic decay function

Simplified logistic decay function

Exponential decay function
}
\description{
List of decay functions

Logistic decay function

Simplified logistic decay function (without constant parameter)

Exponential decay function
}
