% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay_functions.R
\name{decay_functions_period_avg}
\alias{decay_functions_period_avg}
\alias{logistic_decay_period_avg}
\alias{logistic_simple_decay_period_avg}
\alias{exponential_decay_period_avg}
\title{List of period-averaged decay functions (log(1-VE) scale)}
\usage{
decay_functions_period_avg()

logistic_decay_period_avg()

logistic_simple_decay_period_avg()

exponential_decay_period_avg()
}
\value{
Vector of averaged log(1-VE) values for each period
}
\description{
These functions return the mean log(1-VE) over 8-week periods, as used in the toy code.
}
