% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_se_from_ci}
\alias{get_se_from_ci}
\title{Get standard error from confidence interval}
\usage{
get_se_from_ci(ci_low, ci_high)
}
\arguments{
\item{ci_low}{Lower confidence interval}

\item{ci_high}{Upper confidence interval}
}
\value{
Standard error
}
\description{
Get standard error from confidence interval
}
