% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_sample_8_weeks_data.R
\docType{data}
\name{sample_8_weeks_data}
\alias{sample_8_weeks_data}
\title{sample_8_weeks_data}
\format{
A data frame with 12 rows and 5 variables:
\describe{
\item{ date_min }{  Date }
\item{ date_max }{  Date }
\item{ estimate }{  numeric }
\item{ CIlow }{  numeric }
\item{ CIhigh }{  numeric }
}
}
\source{
Source
}
\usage{
sample_8_weeks_data
}
\description{
Description.
}
\keyword{datasets}
